修复了很多BUG

+++ app/src/main/java/cn/com/startai/socket/sign/js/jsInterface/Version.java
@@ -29,7 +29,7 @@ public class Version extends AbsHandlerJsInterface {
         private static final String QUERY_VERSION = "javascript:checkFirmwareVersionResponse('$mac',$update,'$version','$curVersion')";
 
         public static String callJsScmVersion(String mac, boolean update,
-                                              int newVersion, int curVersion) {
+                                              double newVersion, double curVersion) {
             if (mac == null || "".equals(mac)) mac = H5Config.DEFAULT_MAC;
             return QUERY_VERSION.replace("$mac", mac)
                     .replace("$update", String.valueOf(update))
